# regalloc: removes the last use of map for faster allocation